requirements.txt
# MindTrace - Cognitive Pattern Analyzer
# Hyper-Advanced Requirements

# Core
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP
nltk>=3.8
textblob>=0.17
spacy>=3.8

# ML
scikit-learn>=1.3
scipy>=1.11

# Visualization  
plotly>=5.18
wordcloud>=1.9

# Jupyter
ipywidgets>=8.1
ipython>=8.0
jupyter>=1.0

# ===========================================
# HYPER-ADVANCED MODULES
# ===========================================

# GraphQL API
strawberry-graphql>=0.200.0
uvicorn[standard]>=0.24.0

# Redis Caching
redis>=5.0.0

# Apache Spark
pyspark>=3.5.0

# Prometheus Monitoring
prometheus-client>=0.19.0

# Kubernetes
kubernetes>=28.0.0

# LangChain & LLM
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.0

# Vector Databases
pinecone-client>=3.0.0
weaviate-client>=3.25.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Knowledge Graph
networkx>=3.2

# WebSocket API
fastapi>=0.108.0
websockets>=12.0

# GPU Acceleration
cupy-cuda12x>=12.0.0

# Transformers & Deep Learning
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0

# Kafka
kafka-python>=2.0.2

# Cloud Storage
boto3>=1.34.0
google-cloud-storage>=2.14.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Additional Utilities
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# Data Processing
pyarrow>=14.0.0
fastparquet>=2023.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0

# Optional: Jupyter Extensions
ipykernel>=6.26.0
nbconvert>=7.14.0
nbformat>=5.9.0

# Optional: Advanced NLP
gensim>=4.3.0
sentence-transformers>=2.2.0
bertopic>=0.16.0
